django_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3
factory-boy==3.3.0
hypothesis==6.88.1
python-decouple==3.8